pyproject.toml
[project]
name = "viewforge"
version = "0.1.0.dev0"
description = "ViewForge is a minimalist Python UI framework for building desktop apps with web technologies and live reload support."
authors = [{ name = "Israel Dryer", email = "israel.dryer@gmail.com" }]
dependencies = ["pywebview>=4.0", "watchdog"]
requires-python = ">=3.8"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
viewforge = "viewforge.cli:main"
viewforge-reload = "viewforge.cli.reload:run_reload"
